# swagger.yaml
openapi: 3.0.0
info:
  title: LCMS API
  description: Tài liệu API cho hệ thống Quản lý Trung tâm Học tập (LCMS)
  version: 2.0.0
servers:
  - url: http://localhost:3003/api/v1
    description: Local Development Server
security:
  - bearerAuth: []
...
paths:
...
  /parents:
    post:
      tags: [Parent]
      summary: Tạo phụ huynh mới — BO/STAFF (P.1)
      operationId: createParent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateParentRequest'
      responses:
        '201':
          description: Tạo phụ huynh thành công
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Parent'
        '409': { $ref: '#/components/responses/ConflictError' }
        '422': { $ref: '#/components/responses/ValidationError' }
    get:
      tags: [Parent]
      summary: Danh sách phụ huynh — BO/STAFF (P.2)
      operationId: getParents
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
        - name: search
          in: query
          description: Tìm kiếm theo tên phụ huynh
          schema: { type: string }
        - name: isActive
          in: query
          schema: { type: string, enum: ['true', 'false'] }
        - name: studentId
          in: query
          description: Lọc theo học sinh liên kết
          schema: { type: string }
        - name: relationship
          in: query
          schema:
            type: string
            enum: [father, mother, guardian, other]
      responses:
        '200':
          description: Danh sách phụ huynh
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Parent'
                      meta:
                        $ref: '#/components/schemas/PaginationMeta'
  /parents/{id}:
    get:
      tags: [Parent]
      summary: Chi tiết phụ huynh — BO/STAFF (P.3)
      operationId: getParentById
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: Chi tiết phụ huynh
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Parent'
        '404': { $ref: '#/components/responses/NotFoundError' }
    patch:
      tags: [Parent]
      summary: Cập nhật phụ huynh — BO/STAFF (P.4)
      operationId: updateParent
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateParentRequest'
      responses:
        '200':
          description: Cập nhật thành công
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Parent'
        '404': { $ref: '#/components/responses/NotFoundError' }
        '409': { $ref: '#/components/responses/ConflictError' }
    delete:
      tags: [Parent]
      summary: Xóa mềm phụ huynh — BO/STAFF (P.5)
      operationId: deleteParent
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: Xóa thành công
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          deleted: { type: boolean }
                          parentId: { type: string }
        '404': { $ref: '#/components/responses/NotFoundError' }

  # ── Student (Module 3) ────────────────────────────────────────
...
  /student/create-student:
    post:
      tags: [Student]
      summary: Tạo học sinh mới — SO/BO/STAFF (3.1)
      operationId: createStudent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStudentRequest'
      responses:
        '201':
          description: Tạo học sinh thành công
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/User'
        '422': { $ref: '#/components/responses/ValidationError' }
...
  /student/{id}:
    get:
      tags: [Student]
      summary: Chi tiết học sinh (3.3)
      operationId: getStudentById
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: Chi tiết học sinh
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/User'
        '404': { $ref: '#/components/responses/NotFoundError' }

    patch:
      tags: [Student]
      summary: Cập nhật thông tin học sinh (3.5)
      operationId: updateStudent
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStudentRequest'
      responses:
        '200':
          description: Cập nhật thành công
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/User'
        '404': { $ref: '#/components/responses/NotFoundError' }
...
  /reports/branch-dashboard:
    get:
      tags: [Report]
      summary: Dashboard cơ sở — SO/BO (15.1)
      operationId: getBranchDashboard
      parameters:
        - name: branch_id
          in: query
          description: ID cơ sở (SO có thể truyền bất kỳ, BO mặc định là cơ sở của mình)
          schema: { type: string }
      responses:
        '200':
          description: Thống kê tổng quan cơ sở
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          students:
                            type: object
                            properties:
                              total_active: { type: integer }
                              new_this_month: { type: integer }
                              no_class: { type: integer }
                          classes:
                            type: object
                            properties:
                              total_active: { type: integer }
                              total_students_enrolled: { type: integer }
                          finance:
                            type: object
                            properties:
                              revenue_this_month: { type: number }
                              revenue_last_month: { type: number }
                              unpaid_invoices: { type: integer }
                              unpaid_amount: { type: number }
                              overdue_invoices: { type: integer }
                          attendance:
                            type: object
                            properties:
                              this_week_sessions: { type: integer }
                              avg_attendance_rate: { type: number }
  /reports/revenue:
    get:
      tags: [Report]
      summary: Báo cáo doanh thu — SO/BO (15.2)
      operationId: getRevenueReport
      parameters:
        - name: branch_id
          in: query
          schema: { type: string }
        - name: year
          in: query
          required: true
          schema: { type: integer, example: 2025 }
        - name: from_month
          in: query
          schema: { type: integer, minimum: 1, maximum: 12, default: 1 }
        - name: to_month
          in: query
          schema: { type: integer, minimum: 1, maximum: 12, default: 12 }
      responses:
        '200':
          description: Báo cáo doanh thu
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          summary:
                            type: object
                            properties:
                              total_revenue: { type: number }
                              total_invoices: { type: integer }
                              paid_invoices: { type: integer }
                              unpaid_invoices: { type: integer }
                          monthly:
                            type: array
                            items:
                              type: object
                              properties:
                                period: { type: string }
                                revenue: { type: number }
                                invoice_count: { type: integer }
                                paid_count: { type: integer }
                                unpaid_count: { type: integer }
                          by_subject:
                            type: array
                            items:
                              type: object
                              properties:
                                subject_name: { type: string }
                                revenue: { type: number }
  /reports/at-risk-students:
    get:
      tags: [Report]
      summary: Học sinh có nguy cơ bỏ học — SO/BO (15.3)
      operationId: getAtRiskStudents
      parameters:
        - name: branch_id
          in: query
          schema: { type: string }
        - name: consecutive_threshold
          in: query
          description: Số buổi vắng liên tiếp tối thiểu để cảnh báo
          schema: { type: integer, default: 3, minimum: 1 }
      responses:
        '200':
          description: Danh sách học sinh có nguy cơ bỏ học
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          type: object
                          properties:
                            student:
                              type: object
                              properties:
                                _id: { type: string }
                                full_name: { type: string, nullable: true }
                                user_code: { type: string, nullable: true }
                            class:
                              type: object
                              properties:
                                name: { type: string, nullable: true }
                            consecutive_absences: { type: integer }
                            last_present:
                              type: string
                              format: date-time
                              nullable: true
                            parent_phone:
                              type: string
                              nullable: true
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    id:
      name: id
      in: path
      required: true
      schema: { type: string }
...
    page:
      name: page
      in: query
      schema: { type: integer, default: 1, minimum: 1 }
    limit:
      name: limit
      in: query
      schema: { type: integer, default: 20, minimum: 1, maximum: 100 }
  schemas:
...
    ApiResponse:
      type: object
      properties:
        success: { type: boolean, example: true }
        message: { type: string, example: "Thành công" }
        requestId: { type: string, format: uuid }
        timestamp: { type: string, format: date-time }
    ApiErrorResponse:
      type: object
      properties:
        success: { type: boolean, example: false }
        message: { type: string, example: "Lỗi" }
        code: { type: string, example: "NOT_FOUND" }
        requestId: { type: string, format: uuid }
        timestamp: { type: string, format: date-time }
        details: { type: object }
    PaginationMeta:
      type: object
      properties:
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
        total_pages: { type: integer }

    # ── Auth ──────────────────────────────────────────────────────
...
    User:
      type: object
      properties:
        _id: { type: string }
        fullName: { type: string }
        email: { type: string }
        phone: { type: string, nullable: true }
        role:
          type: string
          enum: [SYSTEM_OWNER, BRANCH_OWNER, STAFF, TEACHER, STUDENT, PARENT]
        branchId: { type: string, nullable: true }
        userCode: { type: string }
        avatarUrl: { type: string, nullable: true }
        dateOfBirth: { type: string, format: date-time, nullable: true }
        gender:
          type: string
          enum: [male, female, other]
          nullable: true
        isActive: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    ParentStudentSummary:
      type: object
      properties:
        _id: { type: string }
        fullName: { type: string }
        userCode: { type: string }
        phone: { type: string }
        dateOfBirth: { type: string, format: date-time, nullable: true }
        gender:
          type: string
          enum: [male, female, other]
          nullable: true
        studentInfo:
          type: object
          properties:
            grade: { type: integer, nullable: true }
            schoolName: { type: string, nullable: true }
    ParentInfo:
      type: object
      properties:
        relationship:
          type: string
          enum: [father, mother, guardian, other]
          nullable: true
        studentIds:
          type: array
          items:
            oneOf:
              - type: string
              - $ref: '#/components/schemas/ParentStudentSummary'
    Parent:
      allOf:
        - $ref: '#/components/schemas/User'
        - type: object
          properties:
            dateOfBirth: { type: string, format: date-time, nullable: true }
            gender:
              type: string
              enum: [male, female, other]
              nullable: true
            avatarUrl: { type: string, format: uri, nullable: true }
            parentInfo:
              $ref: '#/components/schemas/ParentInfo'
...
    CreateParentRequest:
      type: object
      required: [fullName, phone]
      properties:
        fullName: { type: string, minLength: 1 }
        email:
          type: string
          format: email
          description: Có thể bỏ trống, hệ thống sẽ tự sinh email nội bộ
        phone: { type: string, minLength: 10 }
        password:
          type: string
          minLength: 8
          description: Nếu bỏ trống, hệ thống dùng mặc định `TempPass@123`
        dateOfBirth: { type: string, format: date }
        gender:
          type: string
          enum: [male, female, other]
        avatarUrl: { type: string, format: uri }
        isActive: { type: boolean }
        parentInfo:
          type: object
          properties:
            relationship:
              type: string
              enum: [father, mother, guardian, other]
            studentIds:
              type: array
              items: { type: string }
    UpdateParentRequest:
      type: object
      properties:
        fullName: { type: string, minLength: 1 }
        email: { type: string, format: email }
        phone: { type: string, minLength: 10 }
        dateOfBirth: { type: string, format: date }
        gender:
          type: string
          enum: [male, female, other]
        avatarUrl: { type: string, format: uri }
        isActive: { type: boolean }
        parentInfo:
          type: object
          properties:
            relationship:
              type: string
              enum: [father, mother, guardian, other]
            studentIds:
              type: array
              items: { type: string }

    # ── Student ───────────────────────────────────────────────────
...
    CreateStudentRequest:
      type: object
      required: [full_name]
      properties:
        full_name: { type: string }
        email: { type: string, format: email }
        phone: { type: string }
        date_of_birth: { type: string, format: date }
        parent_ids:
          type: array
          items: { type: string }
    UpdateStudentRequest:
      type: object
      properties:
        full_name: { type: string }
        phone: { type: string }
        email: { type: string, format: email }
        date_of_birth: { type: string, format: date }

    # ── Class ─────────────────────────────────────────────────────
...
  responses:
...
    NotFoundError:
      description: Không tìm thấy tài nguyên
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    ConflictError:
      description: Dữ liệu đã tồn tại hoặc xung đột trạng thái
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    ValidationError:
      description: Dữ liệu gửi lên không đúng định dạng
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'